Upload de imagens

+++ Back/src/ApiProjeto/Controllers/EventosController.cs
@@ -9,6 +9,8 @@ using ApiProjeto.Application.Contratos;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using ApiProjeto.Application.DTOs;
+using System.IO;
+using Microsoft.AspNetCore.Hosting;
 
 namespace ApiProjeto.Controllers
 {
@@ -17,10 +19,12 @@ namespace ApiProjeto.Controllers
     public class EventosController : ControllerBase
     {
         private readonly IEventoService _eventoService;
+        private readonly IWebHostEnvironment _hostEnvironment;
         
-        public EventosController(IEventoService eventoService)
+        public EventosController(IEventoService eventoService, IWebHostEnvironment hostEnvironment)
         {
             _eventoService = eventoService;
+            _hostEnvironment = hostEnvironment;
         }
 
         [HttpGet]
@@ -71,13 +75,37 @@ namespace ApiProjeto.Controllers
             }
         }
 
+        [HttpPost("upload-image/{eventoId}")]
+        public async Task<IActionResult> UploadImage(int eventoId)
+        {
+            try
+            {
+                var evento = await _eventoService.GetEventoByIdAsync(eventoId, true);
+                if(evento == null) return NoContent();
+            
+                var file = Request.Form.Files[0];
+                if(file.Length > 0)
+                {
+                    DeleteImage(evento.ImagemUrl);
+                    evento.ImagemUrl = await SaveImage(file);
+                }
+                var eventoRetorno = await _eventoService.UpdateEvento(eventoId, evento);
+
+                return Ok(eventoRetorno);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar adicionar evento. Erro: {ex.Message}");
+            }
+        }
+
         [HttpPost]
         public async Task<IActionResult> Post(EventoDTO model)
         {
             try
             {
                 var evento = await _eventoService.AddEvento(model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -108,13 +136,17 @@ namespace ApiProjeto.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento == null) return NoContent();
+
                 if(await _eventoService.DeleteEvento(id))
                 {
+                    DeleteImage(evento.ImagemUrl);
                     return Ok(new {message = "Deletado"});
                 } 
                 else 
                 {
-                    return BadRequest("Evento não deletado");
+                    throw new Exception("Evento não deletado");
                 }
 
             }
@@ -122,6 +154,30 @@ namespace ApiProjeto.Controllers
             {
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar deletar evento. Erro: {ex.Message}");
             }
+        }
+
+        [NonAction]
+        public async Task<string> SaveImage(IFormFile imageFile)
+        {
+            string imageName = new String(Path.GetFileNameWithoutExtension(imageFile.FileName).Take(10).ToArray()).Replace(' ', '-');  //Pega a imagem, apenas os 10 primeiros digitos do nome e da um replace de espçao em branco por traço
+            
+            imageName = $"{imageName}{DateTime.UtcNow.ToString("yymmssfff")}{Path.GetExtension(imageFile.FileName)}";
+            
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/images", imageName);
+
+            using(var fileStream = new FileStream(imagePath, FileMode.Create))
+            {
+                await imageFile.CopyToAsync(fileStream);
+            }
+
+            return imageName;
+        }
+
+        [NonAction]
+        public void DeleteImage(string imageName)
+        {
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/images", imageName);
+            if(System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
         }
     }
 }
